dam-mcp/cloud_functions/document_processor/requirements.txt
functions-framework==3.*
google-cloud-firestore>=2.0.0
cloudevents>=1.0.0
